Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit c2286370e81664f08a025225756523ccc07aff7e, reversing
changes made to 68779e54e1e1c6a731aeab93edef40883edfe76a.

+++ tools/travis/build_byond.sh
@@ -18,7 +18,7 @@ if [ "$BUILD_TOOLS" = false ]; then
     	exit 1
 	fi;
 	if grep 'pixel_[xy] = 0' _maps/**/*.dmm;	then
-    	echo "pixel_x/pixel_y = 0 variables detected in maps, please review to ensure they are not dirty varedits."
+    	echo "pixel_x/pixel_y = 0 variables in maps, please review to ensure they are not dirty varedits."
 	fi;
 	if grep -P '\td[1-2] =' _maps/**/*.dmm;	then
     	echo "d1/d2 cable variables detected in maps, please remove them."
@@ -36,45 +36,44 @@ if [ "$BUILD_TOOLS" = false ]; then
 		echo "Unmanaged global var use detected in code, please use the helpers."
 		exit 1
 	fi;
-#	if grep -i 'centcomm' code/**/*.dm; then
-#		echo "Misspelling(s) of CENTCOM detected in code, please remove the extra M(s)."
-#		exit 1
-#	fi;
-#	if grep -i 'centcomm' _maps/**/*.dmm; then
-#		echo "Misspelling(s) of CENTCOM detected in maps, please remove the extra M(s)."
-#		exit 1
-#	fi;
+	if grep -i 'centcomm' code/**/*.dm; then
+		echo "Misspelling(s) of CENTCOM detected in code, please remove the extra M(s)."
+		exit 1
+	fi;
+	if grep -i 'centcomm' _maps/**/*.dmm; then
+		echo "Misspelling(s) of CENTCOM detected in maps, please remove the extra M(s)."
+		exit 1
+	fi;
 
-    source $HOME/BYOND/byond/bin/byondsetup
+    source $HOME/BYOND-${BYOND_MAJOR}.${BYOND_MINOR}/byond/bin/byondsetup
 	if [ "$BUILD_TESTING" = true ]; then
 		tools/travis/dm.sh -DTRAVISBUILDING -DTRAVISTESTING -DALL_MAPS tgstation.dme
 	else
 		tools/travis/dm.sh -DTRAVISBUILDING tgstation.dme
 		
-		# tools/deploy.sh travis_test someone fix this plz
-		#mkdir travis_test/config
+		#config folder should not be mandatory
+		rm -rf config/*
 
 		#test config
-		cp tools/travis/travis_config.txt config/config.txt #travis_test/
+		cp tools/travis/travis_config.txt config/config.txt
 
 		# get libmariadb, cache it so limmex doesn't get angery
 		if [ -f $HOME/libmariadb ]; then
 			#travis likes to interpret the cache command as it being a file for some reason
 			rm $HOME/libmariadb
+			mkdir $HOME/libmariadb
 		fi
-		mkdir -p $HOME/libmariadb
 		if [ ! -f $HOME/libmariadb/libmariadb.so ]; then
 			wget http://www.byond.com/download/db/mariadb_client-2.0.0-linux.tgz
 			tar -xvf mariadb_client-2.0.0-linux.tgz
 			mv mariadb_client-2.0.0-linux/libmariadb.so $HOME/libmariadb/libmariadb.so
 			rm -rf mariadb_client-2.0.0-linux.tgz mariadb_client-2.0.0-linux
 		fi
+		#violently delete libmariadb.so, because its in the repo for some reason
+		rm -f  libmariadb.so
+    	ln -s $HOME/libmariadb/libmariadb.so libmariadb.so
 	
-		#cd travis_test
-    	#ln -s $HOME/libmariadb/libmariadb.so libmariadb.so
-		DreamDaemon tgstation.dmb -close -trusted -verbose -params "test-run&log-directory=travis"
-		#cd ..
-		#cat data/logs/travis/clean_run.lk #travis_test/
-
+		DreamDaemon tgstation.dmb -close -trusted -params "test-run&log-directory=travis"
+		cat data/logs/travis/clean_run.lk
 	fi;
 fi;
